added email and phone number checks
